[ETCM-129] - starting point of migration

+++ src/main/scala/io/iohk/ethereum/transactions/PendingTransactionsManager.scala
@@ -60,7 +60,7 @@ class PendingTransactionsManager(
   private[this] final val TransactionsPoolSizeGauge =
     metrics.gauge(
       "transactions.pool.size.gauge",
-      () => pendingTransactions.size()
+      () => pendingTransactions.size().toDouble
     )
 
   /**
